pyproject.toml
[project]
name = "catalog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "acatejr@gmail.com", email = "acatejr@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "langchain-community>=0.3.27",
    "lxml>=6.0.0",
    "openai>=1.101.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.7.4",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
catalog = "catalog.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-sugar>=1.0.0",
    "ruff>=0.12.8",
    "mkdocs-material>=9.6.16",
    "rich>=14.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
norecursedirs = ["scratch"]
